+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Budget">
         <activity android:name=".ViewPage" />
+        <activity android:name=".MonthBudget" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
